+++ packages/unlock-js/CHANGELOG.md
@@ -3,6 +3,7 @@
 # 0.39.2
 
 - add `createLock` accepts keyPrice as number
+- `maxNumberOfKeys` is optional when creating a lock. If not passed will use unlimited.
 
 # 0.39.1